Module Odoc_examples.Expansion

Examples of different features of Expansion

For details on what each of the following examples is showing, see the explanations in the Features page

module Simple : sig ... end

Demonstrates simple expansion with a type equality

module Aliases : sig ... end

Demonstrates expansion when a module is an alias to a hidden module.

module DeepConstraint : sig ... end

Demonstrates expansion involving a constraint on a submodule

module DeepConstraint2 : sig ... end

Demonstrates expansion involving a constraint on a submodule, but the submodule is already a simple signature

module TypeSubstitution : sig ... end

module ModuleTypeOf : sig ... end

Demonstrates expanding after recovering the type of a module

module ModuleTypeOfComplications : sig ... end

Demonstrates the interaction of module type of and destructive module substitution