Module Expansion.DeepEquality2

Demonstrates expansion involving an equation on a type in a submodule, but the submodule is already a simple signature

module type MODTYPE = sig ... end

type foo

module M : MODTYPE with type X.t = foo