Module Odoc_examples.Markup

Markup examples.

The OCaml manual gives a comprehensive example of comment placement. This has been replicated in the module Foo below to show how this is rendered by odoc.

module type Foo = sig ... end

The first special comment of the file is the comment associated with the whole module.

module Stop : sig ... end

This module demonstrates the use of stop comments ((**/**))

module Scope : sig ... end

module Preamble_examples : sig ... end

This module demonstrates the various ways that preambles are calculated