Module Scope.A

In this module I can refer to val x declared above as well as type u declared later in the parent module. Elements declared in this signature take priority, so y refers to A.y as opposed to the y declared in the parent signature.

see markup.mli
for a good time

val y : string