Class Foo.my_class

The comment for class my_class

inherit cl

A comment to describe inheritance from cl

val mutable tutu : string

The comment for attribute tutu

val toto : int

The comment for attribute toto.

This comment is not attached to titi since there is a blank line before titi, but is kept as a comment in the class.

val titi : string

method toto : string

Comment for method toto

method m : float -> int

Comment for method m