Deps

The following libraries are dependencies of odoc.

Stdlib

Stdlib The OCaml Standard library.

Fpath

Fpath File system paths, file extensions, path sets and maps.

Cmdliner

Cmdliner Declarative definition of command line interfaces.

Astring

Astring Alternative Char and String modules.

Result

Result Result values.

Yojson

Yojson The Yojson library provides runtime functions for reading and writing JSON data from OCaml. It addresses a few shortcomings of its predecessor json-wheel and is about twice as fast (2.7x reading, 1.3x writing; results may vary). The design goals of Yojson are the following:

Tyxml

Tyxml Typed implementation for HTML, SVG and XML

Biniou

Bi_dump

Bi_inbuf Input buffer

Bi_io Input and output functions for the Biniou serialization format

Bi_outbuf Output buffer

Bi_share [not for general use]

Bi_stream Streaming utilities (experimental)

Bi_util Error handling etc.

Bi_vint Vint: variable-length representation of integers

Odoc-parser

Odoc_parser Parser for ocamldoc formatted comments.

Fmt

Fmt Format pretty-printer combinators.

Fmt_cli Cmdliner support for Fmt.

Fmt_tty Fmt TTY setup.