Module Odoc_examples.Markup

Markup examples

Ocamldoc example from the OCaml manual

The OCaml manual gives a comprehensive example of comment placement. This has been replicated in the module Foo below to show how this is rendered by odoc.

module type Foo = sig ... end

The first special comment of the file is the comment associated with the whole module.

Synopsis, preambles and sections

module Module : sig ... end

This comment is on the declaration of Module as an alias of Hidden__Module

module Module2 : sig ... end

This is the top comment declared in the module Hidden__module2.

module Nonhidden_module : sig ... end

This is the top comment declared in the module Hidden__module2.

module Module3 = Nonhidden_module

This comment is on the declaration of Module3 as an alias of Nonhidden_module

module Nonhidden_module2 : sig ... end

This is the top comment declared in the module Hidden__module2.

module Module4 = Nonhidden_module2

The modules special reference can be used to refer to a list of modules. It uses the synopsis from the modules